+++ src/language-html/print/closing-tag.js
@@ -1,0 +1,175 @@
+"use strict";
+
+const assert = require("assert");
+const {
+  isTextLikeNode,
+  getLastDescendant,
+  isPreLikeNode,
+  hasPrettierIgnore,
+  shouldPreserveContent,
+} = require("../utils.js");
+const {
+  printOpeningTagStartMarker,
+  needsToBorrowNextOpeningTagStartMarker,
+} = require("./opening-tag.js");
+
+function printClosingTag(node, options) {
+  return [
+    node.isSelfClosing ? "" : printClosingTagStart(node, options),
+    printClosingTagEnd(node, options),
+  ];
+}
+
+function printClosingTagStart(node, options) {
+  return node.lastChild &&
+    needsToBorrowParentClosingTagStartMarker(node.lastChild)
+    ? ""
+    : [
+        printClosingTagPrefix(node, options),
+        printClosingTagStartMarker(node, options),
+      ];
+}
+
+function printClosingTagEnd(node, options) {
+  return (
+    node.next
+      ? needsToBorrowPrevClosingTagEndMarker(node.next)
+      : needsToBorrowLastChildClosingTagEndMarker(node.parent)
+  )
+    ? ""
+    : [
+        printClosingTagEndMarker(node, options),
+        printClosingTagSuffix(node, options),
+      ];
+}
+
+function printClosingTagPrefix(node, options) {
+  return needsToBorrowLastChildClosingTagEndMarker(node)
+    ? printClosingTagEndMarker(node.lastChild, options)
+    : "";
+}
+
+function printClosingTagSuffix(node, options) {
+  return needsToBorrowParentClosingTagStartMarker(node)
+    ? printClosingTagStartMarker(node.parent, options)
+    : needsToBorrowNextOpeningTagStartMarker(node)
+    ? printOpeningTagStartMarker(node.next)
+    : "";
+}
+
+function printClosingTagStartMarker(node, options) {
+  assert(!node.isSelfClosing);
+  /* istanbul ignore next */
+  if (shouldNotPrintClosingTag(node, options)) {
+    return "";
+  }
+  switch (node.type) {
+    case "ieConditionalComment":
+      return "<!";
+    case "element":
+      if (node.hasHtmComponentClosingTag) {
+        return "<//";
+      }
+    // fall through
+    default:
+      return `</${node.rawName}`;
+  }
+}
+
+function printClosingTagEndMarker(node, options) {
+  if (shouldNotPrintClosingTag(node, options)) {
+    return "";
+  }
+  switch (node.type) {
+    case "ieConditionalComment":
+    case "ieConditionalEndComment":
+      return "[endif]-->";
+    case "ieConditionalStartComment":
+      return "]><!-->";
+    case "interpolation":
+      return "}}";
+    case "element":
+      if (node.isSelfClosing) {
+        return "/>";
+      }
+    // fall through
+    default:
+      return ">";
+  }
+}
+
+function shouldNotPrintClosingTag(node, options) {
+  return (
+    !node.isSelfClosing &&
+    !node.endSourceSpan &&
+    (hasPrettierIgnore(node) || shouldPreserveContent(node.parent, options))
+  );
+}
+
+function needsToBorrowPrevClosingTagEndMarker(node) {
+  /**
+   *     <p></p
+   *     >123
+   *     ^
+   *
+   *     <p></p
+   *     ><a
+   *     ^
+   */
+  return (
+    node.prev &&
+    node.prev.type !== "docType" &&
+    !isTextLikeNode(node.prev) &&
+    node.isLeadingSpaceSensitive &&
+    !node.hasLeadingSpaces
+  );
+}
+
+function needsToBorrowLastChildClosingTagEndMarker(node) {
+  /**
+   *     <p
+   *       ><a></a
+   *       ></p
+   *       ^
+   *     >
+   */
+  return (
+    node.lastChild &&
+    node.lastChild.isTrailingSpaceSensitive &&
+    !node.lastChild.hasTrailingSpaces &&
+    !isTextLikeNode(getLastDescendant(node.lastChild)) &&
+    !isPreLikeNode(node)
+  );
+}
+
+function needsToBorrowParentClosingTagStartMarker(node) {
+  /**
+   *     <p>
+   *       123</p
+   *          ^^^
+   *     >
+   *
+   *         123</b
+   *       ></a
+   *        ^^^
+   *     >
+   */
+  return (
+    !node.next &&
+    !node.hasTrailingSpaces &&
+    node.isTrailingSpaceSensitive &&
+    isTextLikeNode(getLastDescendant(node))
+  );
+}
+
+module.exports = {
+  printClosingTag,
+  printClosingTagStart,
+  printClosingTagStartMarker,
+  printClosingTagEndMarker,
+  printClosingTagSuffix,
+  printClosingTagEnd,
+  needsToBorrowLastChildClosingTagEndMarker,
+  needsToBorrowParentClosingTagStartMarker,
+  needsToBorrowPrevClosingTagEndMarker,
+};
