+++ src/language-html/utils.js
@@ -11,11 +11,25 @@ const {
   isFrontMatterNode,
 } = require("../common/util.js");
 const {
+  builders: { line, hardline, join },
+  utils: { getDocParts, replaceTextEndOfLine },
+} = require("../document/index.js");
+const {
   CSS_DISPLAY_TAGS,
   CSS_DISPLAY_DEFAULT,
   CSS_WHITE_SPACE_TAGS,
   CSS_WHITE_SPACE_DEFAULT,
 } = require("./constants.evaluate.js");
+const {
+  printOpeningTagEndMarker,
+  needsToBorrowParentOpeningTagEndMarker,
+} = require("./print/opening-tag.js");
+const {
+  printClosingTagStartMarker,
+  printClosingTagEndMarker,
+  needsToBorrowParentClosingTagStartMarker,
+  needsToBorrowLastChildClosingTagEndMarker,
+} = require("./print/closing-tag.js");
 
 const HTML_TAGS = arrayToMap(htmlTagNames);
 const HTML_ELEMENT_ATTRIBUTES = mapObject(htmlElementAttributes, arrayToMap);
@@ -119,20 +133,6 @@ function hasPrettierIgnore(node) {
 
 function isPrettierIgnore(node) {
   return node.type === "comment" && node.value.trim() === "prettier-ignore";
-}
-
-function getPrettierIgnoreAttributeCommentData(value) {
-  const match = value.trim().match(/^prettier-ignore-attribute(?:\s+(.+))?$/s);
-
-  if (!match) {
-    return false;
-  }
-
-  if (!match[1]) {
-    return true;
-  }
-
-  return match[1].split(/\s+/);
 }
 
 /** there's no opening/closing tag or it's considered not breakable */
@@ -602,14 +602,6 @@ function dedentString(text, minIndent = getMinIndentation(text)) {
         .join("\n");
 }
 
-function shouldNotPrintClosingTag(node, options) {
-  return (
-    !node.isSelfClosing &&
-    !node.endSourceSpan &&
-    (hasPrettierIgnore(node) || shouldPreserveContent(node.parent, options))
-  );
-}
-
 function countChars(text, char) {
   let counter = 0;
   for (let i = 0; i < text.length; i++) {
@@ -674,12 +666,44 @@ function isVueSfcBindingsAttribute(attribute, options) {
   );
 }
 
+function getTextValueParts(node, value = node.value) {
+  return node.parent.isWhitespaceSensitive
+    ? node.parent.isIndentationSensitive
+      ? replaceTextEndOfLine(value)
+      : replaceTextEndOfLine(
+          dedentString(htmlTrimPreserveIndentation(value)),
+          hardline
+        )
+    : getDocParts(join(line, splitByHtmlWhitespace(value)));
+}
+
+function getNodeContent(node, options) {
+  let start = node.startSourceSpan.end.offset;
+  if (
+    node.firstChild &&
+    needsToBorrowParentOpeningTagEndMarker(node.firstChild)
+  ) {
+    start -= printOpeningTagEndMarker(node).length;
+  }
+
+  let end = node.endSourceSpan.start.offset;
+  if (
+    node.lastChild &&
+    needsToBorrowParentClosingTagStartMarker(node.lastChild)
+  ) {
+    end += printClosingTagStartMarker(node, options).length;
+  } else if (needsToBorrowLastChildClosingTagEndMarker(node)) {
+    end -= printClosingTagEndMarker(node.lastChild, options).length;
+  }
+
+  return options.originalText.slice(start, end);
+}
+
 module.exports = {
   HTML_ELEMENT_ATTRIBUTES,
   HTML_TAGS,
   htmlTrim,
   htmlTrimPreserveIndentation,
-  splitByHtmlWhitespace,
   hasHtmlWhitespace,
   getLeadingAndTrailingHtmlWhitespace,
   canHaveInterpolation,
@@ -692,7 +716,6 @@ module.exports = {
   getLastDescendant,
   getNodeCssStyleDisplay,
   getNodeCssStyleWhiteSpace,
-  getPrettierIgnoreAttributeCommentData,
   hasPrettierIgnore,
   inferScriptParser,
   isVueCustomBlock,
@@ -710,7 +733,8 @@ module.exports = {
   isUnknownNamespace,
   preferHardlineAsLeadingSpaces,
   preferHardlineAsTrailingSpaces,
-  shouldNotPrintClosingTag,
   shouldPreserveContent,
   unescapeQuoteEntities,
+  getTextValueParts,
+  getNodeContent,
 };
